create thumbnail decoder (sometimes works)
