+++ src/components/ProductCard.tsx
@@ -1,10 +1,13 @@
-const ProductCard = props => {
-const thumbnail = decodeURIComponent(props.thumbnail)
+import { decodeThumbnail } from "../helpers";
 
-return <a href={`https://www.spoonflower.com/en/fabric/${props.slug}`} className="products__list-card secondary" target="_blank">
-    <article>
-    <img src={`https://garden.spoonflower.com/c/${props.designId}/p/f/m/${thumbnail}`} />
-    </article>
-</a>}
+const ProductCard = props => {
+    const thumbnail = decodeThumbnail(props.thumbnail)
+
+    return <a href={`https://www.spoonflower.com/en/fabric/${props.slug}`} className="products__list-card secondary" target="_blank">
+        <article>
+            <img src={`https://garden.spoonflower.com/c/${props.designId}/p/f/m/${thumbnail}`} />
+        </article>
+    </a>
+}
 
 export default ProductCard;
